Updated Chunkio

+++ Code/wwlib/chunkio.h
@@ -57,6 +57,7 @@
 #include "iostruct.h"
 #endif
 
+#include <cstddef>
 #include <type_traits> // std::is_pointer_v
 
 
@@ -150,7 +151,7 @@ public:
 	bool					End_Micro_Chunk();
 
 	// Write data into the file
-	uint32				Write(const void *buf, uint32 nbytes);
+	uint32				Write(const void *buf, size_t nbytes);
 	uint32				Write(const IOVector2Struct & v);
 	uint32				Write(const IOVector3Struct & v);
 	uint32				Write(const IOVector4Struct & v);
@@ -202,7 +203,7 @@ public:
 	uint32				Cur_Micro_Chunk_Length();
 
 	// Read a block of bytes from the output stream.
-	uint32				Read(void *buf, uint32 nbytes);
+	uint32				Read(void *buf, size_t nbytes);
 	uint32				Read(IOVector2Struct * v);
 	uint32				Read(IOVector3Struct * v);
 	uint32				Read(IOVector4Struct * v);
@@ -254,12 +255,12 @@ private:
 */
 #define WRITE_WWSTRING_CHUNK(csave,id,var) { \
 	csave.Begin_Chunk(id); \
-	csave.Write((const char *)var, static_cast<int>(var.Get_Length () + 1)); \
+	csave.Write((const char *)var, static_cast<size_t>(var.Get_Length ()) + 1); \
 	csave.End_Chunk(); }
 
 #define WRITE_WIDESTRING_CHUNK(csave,id,var) { \
 	csave.Begin_Chunk(id); \
-	csave.Write((const wchar_t *)var, static_cast<int>((var.Get_Length () + 1) * 2)); \
+	csave.Write((const wchar_t *)var, (static_cast<size_t>(var.Get_Length ()) + 1) * 2); \
 	csave.End_Chunk(); }
 
 
@@ -323,12 +324,12 @@ private:
 
 #define WRITE_MICRO_CHUNK_WWSTRING(csave,id,var) { \
 	csave.Begin_Micro_Chunk(id); \
-	csave.Write((const char *)var, static_cast<int>(var.Get_Length () + 1)); \
+	csave.Write((const char *)var, static_cast<size_t>(var.Get_Length ()) + 1); \
 	csave.End_Micro_Chunk(); }
 
 #define WRITE_MICRO_CHUNK_WIDESTRING(csave,id,var) { \
 	csave.Begin_Micro_Chunk(id); \
-	csave.Write((const wchar_t *)var, static_cast<int>((var.Get_Length () + 1) * 2)); \
+	csave.Write((const wchar_t *)var, (static_cast<size_t>(var.Get_Length ()) + 1) * 2); \
 	csave.End_Micro_Chunk(); }
